// src/app/api/swagger/models/api-trade-available-operations-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ApiTradeAvailableOperationsDto {
  IsCreateClarificationRequestAllowed?: boolean;
  IsFinishReviewApplicationAllowed?: boolean;
  IsLinkNotTookPlaceTradeAllowed?: boolean;
  IsRepublishAllowed?: boolean;
  IsReturnToReviewApplicationsAllowed?: boolean;
  IsSelectFromParticipantOffersAllowed?: boolean;
  IsSetContractAllowed?: boolean;
  IsSetLapsedTradeAllowed?: boolean;
  IsTradeProtocolDownloadAllowed?: boolean;
}